requirements.txt
requests>=2.32.0
pytz>=2024.1
python-dateutil>=2.9.0
pandas>=2.2.2

